// utils/types.ts
export type Consultation = {
    _id: number;
    patientFirst: string;
    patientLast: string;
    dob: number;
    email: string;
    gender: string;
    location: string;
    file: string;
    number: string;
    healthcare: string;
    consultation: string;
    condition: string;
    consultDate: string;
    consultTime: string;
};